requirements.txt
Flask
psutil
pycryptodome